Refactor - clean code architecture
