+++ Intl.Realty.Firm.Models/Models/ViewModel/RegisterViewModel.cs
@@ -1,8 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc.Rendering;
-using System.ComponentModel.DataAnnotations;
-using System.Security.Cryptography.X509Certificates;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
-namespace DENR_FAPIS.Models
+namespace Intl.Realty.Firm.Models.ViewModel
 {
     public class RegisterViewModel
     {
