+++ Intl.Realty.Firm.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34723.18
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Intl.Realty.Firm", "Intl.Realty.Firm\Intl.Realty.Firm.csproj", "{A64168C2-454D-4F18-81E9-961AB61133BD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Intl.Realty.Firm", "Intl.Realty.Firm\Intl.Realty.Firm.csproj", "{A64168C2-454D-4F18-81E9-961AB61133BD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Intl.Realty.Firm.Models", "Intl.Realty.Firm.Models\Intl.Realty.Firm.Models.csproj", "{651C4B9E-C16F-4F15-A21D-53E434AD5417}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Intl.Realty.Firm.Utility", "Intl.Realty.Firm.Utility\Intl.Realty.Firm.Utility.csproj", "{41247923-7AC6-4EDE-8889-4B1CD707369E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{A64168C2-454D-4F18-81E9-961AB61133BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A64168C2-454D-4F18-81E9-961AB61133BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A64168C2-454D-4F18-81E9-961AB61133BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{651C4B9E-C16F-4F15-A21D-53E434AD5417}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{651C4B9E-C16F-4F15-A21D-53E434AD5417}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{651C4B9E-C16F-4F15-A21D-53E434AD5417}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{651C4B9E-C16F-4F15-A21D-53E434AD5417}.Release|Any CPU.Build.0 = Release|Any CPU
+		{41247923-7AC6-4EDE-8889-4B1CD707369E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{41247923-7AC6-4EDE-8889-4B1CD707369E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{41247923-7AC6-4EDE-8889-4B1CD707369E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{41247923-7AC6-4EDE-8889-4B1CD707369E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
