+++ Intl.Realty.Firm/Controllers/HomeController.cs
@@ -1,5 +1,5 @@
-using Intl.Realty.Firm.Models;
-using Intl.Realty.Firm.ViewModel;
+
+using Intl.Realty.Firm.Models.ViewModel;
 using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics;
