+++ Intl.Realty.Firm/Controllers/TransactionController.cs
@@ -1,4 +1,4 @@
-﻿using Intl.Realty.Firm.ViewModel;
+﻿using Intl.Realty.Firm.Models.ViewModel;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Intl.Realty.Firm.Controllers
