+++ Intl.Realty.Firm/Intl.Realty.Firm.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -17,6 +17,11 @@
 
   <ItemGroup>
     <Folder Include="wwwroot\files\img\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Intl.Realty.Firm.Models\Intl.Realty.Firm.Models.csproj" />
+    <ProjectReference Include="..\Intl.Realty.Firm.Utility\Intl.Realty.Firm.Utility.csproj" />
   </ItemGroup>
 
 </Project>
