+++ Intl.Realty.Firm/Repository/IRepository/IRepository.cs
@@ -1,0 +1,14 @@
+﻿using System.Linq.Expressions;
+
+namespace Intl.Realty.Firm.Repository.IRepository
+{
+    public interface IRepository<T> where T : class
+    {
+        //T - Category
+        IEnumerable<T> GetAll(Expression<Func<T, bool>>? filter = null, string? includeProperties = null);
+        T Get(Expression<Func<T, bool>> filter, string? includeProperties = null, bool tracked = false);
+        void Add(T entity);
+        void Remove(T entity);
+        void RemoveRange(IEnumerable<T> entity);
+    }
+}
