+++ Intl.Realty.Firm/Repository/IRepository/IUnitOfWork.cs
@@ -1,0 +1,7 @@
+﻿namespace Intl.Realty.Firm.Repository.IRepository
+{
+    public interface IUnitOfWork
+    {
+        void Save();
+    }
+}
