+++ Intl.Realty.Firm/Views/Account/SignUp.cshtml
@@ -1,4 +1,4 @@
-﻿@using DENR_FAPIS.Models;
+﻿@using Intl.Realty.Firm.Models.ViewModel
 
 @model RegisterViewModel
