+++ Intl.Realty.Firm/Views/Shared/Error.cshtml
@@ -1,4 +1,5 @@
-﻿@model ErrorViewModel
+﻿@using Intl.Realty.Firm.Models.ViewModel
+@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
