+++ Intl.Realty.Firm/Views/Shared/_LayoutAfterLogin.cshtml
@@ -1,4 +1,5 @@
-﻿@using Intl.Realty.Firm.ViewModel
+﻿@using Intl.Realty.Firm.Models.ViewModel
+
 <html lang="en">
 <head>
     <meta charset="utf-8" />
